feat(picker): Add support for mixed grouped and not grouped pickers
Previously, a grouped multi picker required all primary options to be categories to group together the final choices. Now, there will be a new picker which will support both final choices and categories in the primary list of options, and will show the secondary options if one of the categories is selected.